server/requirements.txt
<<<<<<< HEAD
asgiref==3.3.4
coverage==5.5
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
gunicorn==20.1.0
drf-firebase-auth==1.0.0
<<<<<<< HEAD
 
=======
asgiref==3.3.4
CacheControl==0.12.6
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
coverage==5.5
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
drf-firebase-auth==1.0.0
firebase-admin==4.5.3
google-api-core==1.31.3
google-api-python-client==2.24.0
google-auth==2.2.1
google-auth-httplib2==0.1.0
google-cloud-core==2.1.0
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
Markdown==3.3.4
msgpack==1.0.2
packaging==21.0
proto-plus==1.19.2
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.7
>>>>>>> 2b9ae1411f3a179dba6d78ebaef6241f3786bcc7
=======
Pillow
>>>>>>> f9ff65b2c29c2e534295c659d56c03502f7be1ee
